preview game custom artwork before update
