+++ frontend/src/components/Details/Cover.vue
@@ -1,34 +1,19 @@
 <script setup lang="ts">
 import storeDownload from "@/stores/download";
-import type { Rom } from "@/stores/roms";
-import { useTheme } from "vuetify";
-const theme = useTheme();
 
-defineProps<{ rom: Rom }>();
+defineProps<{ romId: number; src: string; lazySrc: string }>();
 const downloadStore = storeDownload();
 </script>
 <template>
   <v-card
     elevation="2"
-    :loading="downloadStore.value.includes(rom.id) ? 'romm-accent-1' : false"
+    :loading="downloadStore.value.includes(romId) ? 'romm-accent-1' : false"
   >
     <v-img
-      :value="rom.id"
-      :key="rom.id"
-      :src="
-        !rom.igdb_id && !rom.has_cover
-          ? `/assets/default/cover/big_${theme.global.name.value}_unmatched.png`
-          : !rom.has_cover
-          ? `/assets/default/cover/big_${theme.global.name.value}_missing_cover.png`
-          : `/assets/romm/resources/${rom.path_cover_l}`
-      "
-      :lazy-src="
-        !rom.igdb_id && !rom.has_cover
-          ? `/assets/default/cover/small_${theme.global.name.value}_unmatched.png`
-          : !rom.has_cover
-          ? `/assets/default/cover/small_${theme.global.name.value}_missing_cover.png`
-          : `/assets/romm/resources/${rom.path_cover_s}`
-      "
+      :value="romId"
+      :key="romId"
+      :src="src"
+      :lazy-src="lazySrc"
       :aspect-ratio="3 / 4"
     >
       <slot name="editable"></slot>
