+++ frontend/src/components/Dialog/Rom/EditRom.vue
@@ -5,9 +5,10 @@ import storeRoms from "@/stores/roms";
 import type { Events } from "@/types/emitter";
 import type { Emitter } from "mitt";
 import { inject, ref } from "vue";
-import { useDisplay } from "vuetify";
+import { useDisplay, useTheme } from "vuetify";
 
 // Props
+const theme = useTheme();
 const { xs, mdAndDown, smAndDown, md, lgAndUp } = useDisplay();
 const show = ref(false);
 const rom = ref<UpdateRom>();
@@ -91,6 +92,7 @@ async function updateRom() {
 
 function closeDialog() {
   show.value = false;
+  imagePreviewUrl.value = "";
 }
 </script>
 
@@ -174,7 +176,25 @@ function closeDialog() {
           <v-col cols="12" md="4" lg="3">
             <cover
               :class="{ 'mx-16': smAndDown, 'ml-2': md, 'my-4': smAndDown }"
-              :rom="rom"
+              :romId="rom.id"
+              :src="
+                imagePreviewUrl
+                  ? imagePreviewUrl
+                  : !rom.igdb_id && !rom.has_cover
+                  ? `/assets/default/cover/big_${theme.global.name.value}_unmatched.png`
+                  : !rom.has_cover
+                  ? `/assets/default/cover/big_${theme.global.name.value}_missing_cover.png`
+                  : `/assets/romm/resources/${rom.path_cover_l}`
+              "
+              :lazy-src="
+                imagePreviewUrl
+                  ? imagePreviewUrl
+                  : !rom.igdb_id && !rom.has_cover
+                  ? `/assets/default/cover/small_${theme.global.name.value}_unmatched.png`
+                  : !rom.has_cover
+                  ? `/assets/default/cover/small_${theme.global.name.value}_missing_cover.png`
+                  : `/assets/romm/resources/${rom.path_cover_s}`
+              "
             >
               <template v-slot:editable>
                 <v-chip-group class="position-absolute edit-cover pa-0">
